Continuing to pull on the thread of noms support
